refactor(workflow): split comms.rs into routing submodules
